Move NoiseLearner results to results
