+++ qiskit_ibm_runtime/utils/noise_learner_result_decoder.py
@@ -16,7 +16,7 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
-from .noise_learner_result import LayerError, NoiseLearnerResult, PauliLindbladError
+from ..results.noise_learner import LayerError, NoiseLearnerResult, PauliLindbladError
 from .result_decoder import ResultDecoder
 
 if TYPE_CHECKING:
