+++ test/unit/test_noise_learner_result.py
@@ -20,7 +20,7 @@ from qiskit.quantum_info import PauliList
 from qiskit_aer import AerSimulator
 
 from qiskit_ibm_runtime.fake_provider import FakeKyiv
-from qiskit_ibm_runtime.utils.noise_learner_result import PauliLindbladError, LayerError
+from qiskit_ibm_runtime.results.noise_learner import PauliLindbladError, LayerError
 
 from ..ibm_test_case import IBMTestCase
