+++ test/unit/visualization/test_draw_layer_error.py
@@ -17,7 +17,7 @@ from qiskit.quantum_info import PauliList
 from qiskit_aer import AerSimulator
 
 from qiskit_ibm_runtime.fake_provider import FakeKyiv
-from qiskit_ibm_runtime.utils.noise_learner_result import PauliLindbladError, LayerError
+from qiskit_ibm_runtime.results.noise_learner import PauliLindbladError, LayerError
 from qiskit_ibm_runtime.visualization import draw_layer_error_map, draw_layer_errors_swarm
 
 from ...ibm_test_case import IBMTestCase
